fix material ui and improve folder structure
