--- pages/_app.js
@@ -1,37 +1,0 @@
-import App from 'next/app';
-import Head from 'next/head';
-import { appWithTranslation } from 'i18n';
-
-import '../styles/globals.css';
-import { createMuiTheme } from '@material-ui/core/styles';
-import { ThemeProvider } from '@material-ui/styles';
-
-const theme = createMuiTheme({
-  palette: {
-    primary: {
-      main: '#212121',
-    },
-    secondary: {
-      main: '#D32822',
-    },
-  },
-});
-
-const MyApp = ({ Component, pageProps }) => (
-  <>
-    <Head>
-      <title>Nuse Gym</title>
-      <link rel="icon" href="/favicon.ico" />
-      <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
-      <meta name="viewport" content="minimum-scale=1, initial-scale=1, width=device-width" />
-    </Head>
-
-    <ThemeProvider theme={theme}>
-      <Component {...pageProps} />
-    </ThemeProvider>
-  </>
-);
-
-MyApp.getInitialProps = async (appContext) => ({ ...(await App.getInitialProps(appContext)) });
-
-export default appWithTranslation(MyApp);
