+++ src/pages/login.js
@@ -1,9 +1,9 @@
 import PropTypes from 'prop-types';
-import { withTranslation } from 'i18n';
+import { withTranslation } from 'src/i18n';
 
 import { Typography } from '@material-ui/core';
 
-import PageTitle from 'components/util/PageTitle';
+import PageTitle from 'src/components/util/PageTitle';
 
 const Login = ({ t }) => (
   <>
