+++ src/theme.js
@@ -1,0 +1,14 @@
+import { createMuiTheme } from '@material-ui/core/styles';
+
+const theme = createMuiTheme({
+  palette: {
+    primary: {
+      main: '#212121',
+    },
+    secondary: {
+      main: '#D32822',
+    },
+  },
+});
+
+export default theme;
